requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
joblib>=1.3
matplotlib>=3.8
